cms completed

+++ app/Http/Controllers/CategoryController.php
@@ -16,6 +16,7 @@ class CategoryController extends Controller
      */
     public function index()
     {
+        // dd(Category::first()->posts());
         return view('category.index')->with('categories', Category::latest()->get());
     }
 
@@ -98,6 +99,11 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->posts->count() > 0) {
+            session()->flash('error', 'Category can\'t be deleted because it has some posts!.');
+            return redirect()->back();
+        }
+
         $category->delete();
 
         session()->flash('success', 'Category deleted successfully.');
